requirements.txt
azure-storage-blob
azure-storage-queue==12.12.0
browserbase==1.2.0
celery==5.5.1
celery-redbeat==2.3.2
cryptography==42.0.5
docker==7.1.0
flask-cors
Flask-SQLAlchemy==3.1.1
Flask==3.0.2
langchain
langchain-openai
pillow==11.1.0
playwright
psycopg2-binary
markdownify==0.11.6
PyJWT==2.8.0
python-dotenv
redis
sentry-sdk==2.12.0
uv==0.6.3
kubernetes